[CHG] some refactoring / loaders etc.
