+++ src/utils/CountryNames.js
@@ -1,8 +1,7 @@
 import { invertObj, prop, toLower, propOr } from "ramda";
-import { isoCountries } from "../constants/countries";
+import { isoCountries } from "../constants/";
 
 const countriesIso = invertObj(isoCountries);
-
 
 export const getCountryName = (countryCode) =>
   prop(countryCode, isoCountries);
